+++ lib/src/delta_listenable.dart
@@ -1,0 +1,6 @@
+typedef DeltaHandler<T> = void Function(T);
+
+abstract class DeltaListenable<T> {
+  void addListener(DeltaHandler<T> listener);
+  void removeListener(DeltaHandler<T> listener);
+}