+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## [0.0.1+1] - Fix package link
 
 * Fixes the url on pub.dev to the Github repo
+
+## [0.0.2] - Add beforeBuild handler and example
+
+* Adds the beforeBuild handler, called before setState if set.
+* Add basic example.
+* Restructure files